Factor out the test runner loop
