+++ duktape.js
@@ -12,10 +12,6 @@ var fs = require('fs');
 var child_process = require('child_process');
 var runner_support = require('./runner_support');
 
-var testCount = 0;
-var testSuccess = 0;
-var testOutOfDate = 0;
-
 var dukCommand = './duk';
 
 // Key for .res (e.g. test.res.duktape2_0), automatic based on Duktape.version.
@@ -32,91 +28,23 @@ var dukKey = (function () {
 })();
 console.log('Duktape result key is: test.res.' + dukKey);
 
-// List of keys for inheriting results from previous versions.
-var dukKeyList = runner_support.keyList(dukKey, 'Duktape');
-console.log('Duktape key list for inheriting results is:', dukKeyList);
+function runTest(evalcode) {
+    var script = 'var evalcode = ' + JSON.stringify(evalcode) + ';\n' +
+                 'try {\n' +
+                 '    var res = eval(evalcode);\n' +
+                 '    if (res !== true && res !== 1) { throw new Error("failed: " + res); }\n' +
+                 '    console.log("[SUCCESS]");\n' +
+                 '} catch (e) {\n' +
+                 '    console.log("[FAILURE]", e);\n' +
+                 '    /*throw e;*/\n' +
+                 '}\n';
 
-// Run test / subtests, recursively.  Report results, indicate data files
-// which are out of date.
-function runTest(parents, test, sublevel) {
-    var testPath = parents.join(' -> ') + ' -> ' + test.name;
-
-    if (typeof test.exec === 'function') {
-        var src = test.exec.toString();
-        var m = /^function\s*\w*\s*\(.*?\)\s*\{\s*\/\*([\s\S]*?)\*\/\s*\}$/m.exec(src);
-        var evalcode;
-        if (m) {
-            evalcode = '(function test() {' + m[1] + '})();';
-        } else {
-            evalcode = '(' + src + ')()';
-        }
-        //console.log(evalcode);
-
-        var script = 'var evalcode = ' + JSON.stringify(evalcode) + ';\n' +
-                     'try {\n' +
-                     '    var res = eval(evalcode);\n' +
-                     '    if (res !== true && res !== 1) { throw new Error("failed: " + res); }\n' +
-                     '    console.log("[SUCCESS]");\n' +
-                     '} catch (e) {\n' +
-                     '    console.log("[FAILURE]", e);\n' +
-                     '    /*throw e;*/\n' +
-                     '}\n';
-
-        fs.writeFileSync('duktest.js', script);
-        var stdout = child_process.execFileSync(dukCommand, [ 'duktest.js' ], {
-            encoding: 'utf-8'
-        });
-        //console.log(stdout);
-
-        var success = false;
-        if (/^\[SUCCESS\]$/gm.test(stdout)) {
-            success = true;
-            testSuccess++;
-        } else {
-            //console.log(stdout);
-        }
-        testCount++;
-
-        if (test.res) {
-            // Take expected result from newest Duktape version not newer
-            // than current version.
-            var expect = void 0;
-            dukKeyList.forEach(function (k) {
-                if (test.res[k] !== void 0) {
-                    expect = test.res[k];
-                }
-            });
-
-            if (expect === success) {
-                // Matches.
-            } else if (expect === void 0 && !success) {
-                testOutOfDate++;
-                console.log(testPath + ': test result missing, res: ' + expect + ', actual: ' + success);
-            } else {
-                testOutOfDate++;
-                console.log(testPath + ': test result out of date, res: ' + expect + ', actual: ' + success);
-            }
-        } else {
-            testOutOfDate++;
-            console.log(testPath + ': test.res missing');
-        }
-    }
-    if (test.subtests) {
-        var newParents = parents.slice(0);
-        newParents.push(test.name);
-        test.subtests.forEach(function (v) { runTest(newParents, v, sublevel + 1); });
-    }
-}
-
-for (var suitename in runner_support.suites) {
-    var testsuite = runner_support.suites[suitename];
-    console.log('');
-    console.log('**** ' + suitename + ' ****');
-    console.log('');
-    testsuite.tests.forEach(function (test) {
-        runTest([ suitename ], test);
+    fs.writeFileSync('duktest.js', script);
+    var stdout = child_process.execFileSync(dukCommand, [ 'duktest.js' ], {
+        encoding: 'utf-8'
     });
+
+    return /^\[SUCCESS\]$/gm.test(stdout);
 }
 
-console.log(testCount + ' tests executed: ' + testSuccess + ' success, ' + (testCount - testSuccess) + ' fail');
-console.log(testOutOfDate + ' tests are out of date (data-*.js file .res)');
+runner_support.runTests(runTest, dukKey, 'Duktape');
